+++ packages/fiori-docs-embeddings/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @sap-ux/fiori-docs-embeddings
 
+## 0.4.3
+
+### Patch Changes
+
+-   a990ed6: feat: add @sap-ux/create/README.md as resource
+
+## 0.4.2
+
+### Patch Changes
+
+-   cfe9c13: Add deep link to package and changelog to README.md
+
 ## 0.4.1
 
 ### Patch Changes